+++ EssentialFeed/EssentialFeed/Feed Cache/LocalFeedLoader.swift
@@ -7,23 +7,6 @@
 
 import Foundation
 
-private final class FeedCachePolicy {
-    private init() {}
-    
-    private static let calendar = Calendar(identifier: .gregorian)
-
-    private static var maxCacheAgeInDays: Int {
-        return 7
-    }
-
-    static func validate(_ timestamp: Date, against date: Date) -> Bool {
-        guard let maxCacheAge = calendar.date(byAdding: .day, value: maxCacheAgeInDays, to: timestamp) else {
-            return false
-        }
-        return date < maxCacheAge
-    }
-}
-
 public final class LocalFeedLoader {
     private let store: FeedStore
     private let currentDate: () -> Date
